finish_AddPickupAndDelivery

+++ src/sharing/dao/node.h
@@ -17,8 +17,8 @@ enum class NodeType : int { VEHICLE_START = 1, VEHICLE_END = 2, ORDER_DIRECT = 3
 class Node {
  public:
   Node() {}
-  Node(const Order& order, const std::int64_t &system_index, const NodeType &nodetype);
-  Node(const Vehicle& vehicle, const std::int64_t &system_index, const NodeType &nodetype);
+  Node(const Order& order, const std::size_t &system_index, const NodeType &nodetype);
+  Node(const Vehicle& vehicle, const std::size_t &system_index, const NodeType &nodetype);
 
   ~Node() {}
   auto ToJson() -> json;
@@ -26,15 +26,15 @@ class Node {
   NodeType nodetype = NodeType::UNKNOW;
   Order order = DummyOrder;
   Vehicle vehicle = DummyVehicle;
-  std::int64_t system_index = -1;
+  std::size_t system_index = -1;
 };
 
-Node::Node(const Order& order, const std::int64_t &system_index, const NodeType &nodetype) {
+Node::Node(const Order& order, const std::size_t &system_index, const NodeType &nodetype) {
   this->order = order;
   this->system_index = system_index;
   this->nodetype = nodetype;
 }
-Node::Node(const Vehicle& vehicle, const std::int64_t &system_index, const NodeType &nodetype) {
+Node::Node(const Vehicle& vehicle, const std::size_t &system_index, const NodeType &nodetype) {
   this->vehicle = vehicle;
   this->system_index = system_index;
   this->nodetype = nodetype;
